refactor: Move `install_plugins` to `meltano.core.plugin_install_service`
This lets us eliminate one dependency on `meltano.cli` in `meltano.core`, though there's still
https://github.com/meltano/meltano/blob/4efb91d68bb8e5c2364ea7ae2f032a937959d3f2/src/meltano/core/tracking/tracker.py#L520-L569

which I'm not sure yet how to decouple.
